refactor :: 구글맵 버그 수정

+++ app/src/main/java/com/uiel/scul/model/culture/CultureDetailResponse.kt
@@ -18,6 +18,6 @@ data class CultureDetailResponse(
     val serviceStartDate: String,
     val serviceEndDate: String,
     val cultureLink: String,
-    val ycoordinate: Float,
-    val xcoordinate: Float,
+    val ycoordinate: Double,
+    val xcoordinate: Double,
 )
